feature: Enhance LLM integration with BedrockChat and AgentRunner (#16)
### Description of changes

This pull request introduces several updates to improve LLM integration:
- Added system prompt support and maxTokens validation in `BedrockChat`
and `LLMAgent`.
- Updated `ChatMessage` and `ChatRequest` to handle multiple messages
effectively.
- Implemented `BedrockChat` and `AgentRunner` classes for seamless
interaction with large language models.

### Checklist

- [ ] Tests were added or updated to reflect the changes
- [ ] Documentation was updated where necessary
- [ ] All commits contain clear descriptions of the changes made

### Additional Notes

N/A

+++ ai-core/src/main/scala/wvlet/ai/agent/ModelConfig.scala
@@ -22,7 +22,7 @@ case class ModelConfig(
     /**
       * The maximum number of tokens allowed in the generated response.
       */
-    maxOutputTokens: Option[Long] = None,
+    maxOutputTokens: Option[Int] = None,
 
     /**
       * A list of strings that, if encountered one of them, will stop the generation process. None
@@ -42,10 +42,10 @@ case class ModelConfig(
     reasoningConfig: Option[ReasoningConfig] = None
 ):
 
-  def withTemperature(t: Double): ModelConfig     = this.copy(temperature = Some(t))
-  def withTopP(p: Double): ModelConfig            = this.copy(topP = Some(p))
-  def withTopK(k: Int): ModelConfig               = this.copy(topK = Some(k))
-  def withMaxOutputTokens(max: Long): ModelConfig = this.copy(maxOutputTokens = Some(max))
+  def withTemperature(t: Double): ModelConfig    = this.copy(temperature = Some(t))
+  def withTopP(p: Double): ModelConfig           = this.copy(topP = Some(p))
+  def withTopK(k: Int): ModelConfig              = this.copy(topK = Some(k))
+  def withMaxOutputTokens(max: Int): ModelConfig = this.copy(maxOutputTokens = Some(max))
   def withStopSequences(sequences: List[String]): ModelConfig = this.copy(stopSequences =
     Some(sequences)
   )
@@ -90,5 +90,5 @@ case class ReasoningConfig(
     /**
       * The maximum number of tokens allowed for reasoning.
       */
-    reasoningBudget: Option[Long] = None
+    reasoningBudget: Option[Int] = None
 )
